split out prices and products into routes
